requirements.txt
agno
dotenv
dash-mantine-components==0.12.1
dash==2.15.0
plotly==5.18.0
leia-br